TComOutputBitstream: switch to byte oriented buffering
The introduction of the vector maintained the existing behaviour that
worked in four-byte words.  Switch to inserting bytes into the fifo in
single byte words.

This commit does not change the size of the intermediate scratch buffer
that is four bytes long.

From: David Flynn <davidf@rd.bbc.co.uk>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-3.0-dev@872 ce97f6c2-8f86-0410-b44d-af84b13ec34f