encoder: refactor creation of output bitstreams
Previously, the encoder used a circular buffer and attempted to re-use old
bitstream buffers.  Each bitstream buffer corresponded to a single access
unit.  The upcoming NALunit refactor requires that each access unit is
actually a list of buffers, one for each NAL unit.  In such a situation,
it makes little sense to try and recycle the old storage.

This commit changes the life cycle of the buffers so that:
 - New buffers are created inside the GOP encoder, one for each accessunit
 - They are appended to an output list
 - The top level of the encoder application destroys the buffer after it
   has written them out.

From: David Flynn <davidf@rd.bbc.co.uk>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-3.0-dev@882 ce97f6c2-8f86-0410-b44d-af84b13ec34f