NALunits: introduce new NALUnit representation, writing and rbsp conversion
This commit introduces a NALUnit structure, representing the values
in a NAL unit.  Each NAL unit also contains a bitstream object for
the RBSPayload.

Every NAL unit the encoder creates should have its bitstream created
by the NALUnit constructor.

A function is provided to add rbsp_trailing_bits.  This should be
used in preference to doing it manually.  This function does not
insert cabac_zero_word's.

A function is provided to write an entire NALUnit to an std::ostream
(ie, a file, buffer, etc).  This function performs the start code
anti-emulation previously performed by TCom[Output]Bitstream.

A function is provided to write out NALunits in a manner consistent with
AVC AnnexB.  This function performs the correct insertion of zero_byte's
according to the NALunit type[1].  There is no need to manually write out
startcodes anywhere else.
This function is specific to the encoder and could not be called by
TVideoIOBitsStartCode.  To make this work, the std::fstream in
TVideoIOBitsStartCode has been made public.  A future commit will delete
TVideoIOBitsStartCode completely, since it only serves as a wrapper
around std::fstream, something that can be done in less code by the
top level encoder application.

SPS/PPS/SEI and Slice NALunit writing has been updated to use the above
changes.  In respect to slice writing, the use of independent NAL units
removes the requirement for trying to track the start positions of
each slice for use by the anti-emulation code.

The per-picture bit count output of the decoder now represents the
number of NALUnit header bytes plus RBSP bytes in VCL NALunits in the
current access unit.  It does not include emulation prevention bytes,
a subsequent patch will update the stats output to be more fine grained

[1] Currently, the decoder can't handle a lack of a zero_byte.  A hack
    is implemented to work around this by always inserting the
    zero_byte.

From: David Flynn <davidf@rd.bbc.co.uk>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-3.0-dev@883 ce97f6c2-8f86-0410-b44d-af84b13ec34f