encoder: fix handling of end of source video file (closes #795)
It is only possible to identify end-of-file when trying to read past
the last byte of an file.  Consequently, the encoder has to be capable
of being informed that it has already seen the last picture.

Update the encoder to handle this condition, allows a NULL picture to
be inserted that has the flush option set to force the encoding of a
partial GOP.

The GOP structure is currently built using information about the number
of pictures that will be encoded in the entire sequence.  Given that
the sequence that has encountered EOF is shorter than expected, a hack
is required to update the sequence length information.  Ideally, the
core encoder should not be aware of the sequence length.

From: David Flynn <dflynn@rim.com>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-8.2-dev@2918 ce97f6c2-8f86-0410-b44d-af84b13ec34f