encoder: enable control of profile and level signalling
Adds three new options to the encoder:
  --Profile=none|main|main10|main-still-picture
  --Level=none|1|2|2.1|3|3.1|4|4.1|5|5.1|5.2|6|6.1|6.2
  --Tier=main|high

Defaults are Profile=none, Level=none, Tier=main

If main10 is selected and the encoder is configured to be in 8-bit, the
compatibility bit with main profile will be set.

There is currently no checking to verify that any options are within the
constraints of a given level or profile.

From: David Flynn <dflynn@rim.com>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-8.2-dev@2932 ce97f6c2-8f86-0410-b44d-af84b13ec34f