RExt: fix distortion calculation overflow with -DFULL_NBIT=1
When bitdepths is greater than 8, the use of Int types can overflow
when computing distortion calculations (particularly for large blocks).
This can lead to peculiar results on high-bitdepth sequences where
a frame may suddenly have an absurdly low PSNR.

NB, when FULL_NBIT=0, the distortion calculation is performed as if
it were 8bit data.

This patch introduces a Distortion type that is used for all distortion
calculations.  When FULL_NBIT=1, the Distortion type equivalent to
Int64, otherwise the type remains as Int.

From: Karl Sharman <karl.sharman@eu.sony.com>

git-svn-id: https://hevc.hhi.fraunhofer.de/svn/svn_HEVCSoftware/branches/HM-range-extensions@3628 ce97f6c2-8f86-0410-b44d-af84b13ec34f